NEL benchmark: refactoring of data class handling.
