+++ benchmarks/nel/configs/nel.cfg
@@ -13,7 +13,7 @@ gpu_allocator = null
 
 [nlp]
 lang = "en"
-pipeline = ["sentencizer","entity_ruler","ner","entity_linker"]
+pipeline = ["sentencizer","ner","entity_linker"]
 disabled = []
 before_creation = null
 after_creation = null
@@ -22,13 +22,36 @@ tokenizer = {"@tokenizers":"spacy.Tokenizer.v1"}
 
 [components]
 
+;[components.senter]
+;factory = "senter"
+;overwrite = false
+;scorer = {"@scorers":"spacy.senter_scorer.v1"}
+;
+;[components.senter.model]
+;@architectures = "spacy.Tagger.v2"
+;nO = null
+;normalize = false
+;
+;[components.senter.model.tok2vec]
+;@architectures = "spacy.Tok2Vec.v2"
+;
+;[components.senter.model.tok2vec.embed]
+;@architectures = "spacy.MultiHashEmbed.v2"
+;width = 16
+;attrs = ["NORM","PREFIX","SUFFIX","SHAPE","SPACY"]
+;rows = [1000,500,500,500,50]
+;include_static_vectors = true
+;
+;[components.senter.model.tok2vec.encode]
+;@architectures = "spacy.MaxoutWindowEncoder.v2"
+;width = 16
+;depth = 2
+;window_size = 1
+;maxout_pieces = 2
+
 [components.sentencizer]
 factory = "sentencizer"
 punct_chars = null
-
-[components.entity_ruler]
-source = "${paths.base_nlp}"
-component = "entity_ruler"
 
 [components.ner]
 source = "${paths.base_nlp}"
@@ -71,7 +94,6 @@ lookups = null
 kb_path = ${paths.kb}
 
 [initialize.tokenizer]
-
 
 [corpora]
