+++ benchmarks/nel/requirements.txt
@@ -1,1 +1,3 @@
 pyyaml
+requests
+tqdm
